requirements.txt
pynput
sounddevice
numpy
scipy
google-generativeai
python-dotenv
pyautogui
